fix(scheduler): stabilize round timing and state transition
Track aggregated/applied updates so ApplyUpdate only follows broadcast and
workers don’t churn; reset aggregation flags on
broadcast and error paths. Normalize short/long timeouts for ~100ms RTT,
align PS action retries with scheduler cadence.
Fix training sleep precision to avoid truncation delays.

+++ executors/accelerate/src/hypha/accelerate_executor/training.py
@@ -82,7 +82,7 @@ def system_time_to_epoch_ms(timeout: object) -> int | None:
 
 
 def sleep_until_epoch_ms(target_ms: int) -> None:
-    now_ms = int(time.time() * 1000.0)
+    now_ms = time.time() * 1000.0
     if target_ms > now_ms:
         time.sleep((target_ms - now_ms) / 1000.0)
